backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
google-genai>=0.3.0
python-dotenv>=1.0.0
websockets>=12.0
pydantic>=2.7.0
pypdf>=4.0.0
python-multipart>=0.0.9
google-auth>=2.0.0
groq>=0.5.0
pytest>=8.0.0
httpx>=0.27.0
